Add support for marker shapes, and fix so that colors can be updated from GUI
The shapes are stored in a 512x512 texture atlas, markershapes.png, that can be
generated from misc/markershapes.svg. So it should be easy to change them or
add more shapes if needed.

Anti-aliasing (multisampling) is now also disabled for the WebGL context,
since marker shapes are rendered with alpha blending. Disable anti-aliasing
should improve the performance on high-DPI displays, where multisampling
could be expensive and a bit wasteful.