Replace fixed color scale with texture
This texture will either store a selected color scale from D3-interpolate, or a
default rainbow scale if no color scale or "ownColorFromColumn" is selected.
There was also a bug in the property normalization in the vertex shader, which
is now fixed so that it works event when the min-value is not 0.0.