Increase LUT size to 4096 and add support for using colors from marker data
Color data is now stored in the last component of each marker vertex. For gene
expression markers, color from the data is used if we set
markerUtils._uniqueColor to true and markerUtils._uniqueColorSelector to the
name of the column that contains a hex color. For cell morphology markers, the
current property column is interpreted as hex colors if 'ownColorFromColumn' is
selected as colorscale.