Store marker data in struct-of-arrays (SoA) format
Arranging the data as struct-of-arrays (SoA) instead of array-of-structs (AoS) should avoid duplication of keys from the column headers, and could further allow some of the optimizations described in [1] to be performed by the browser. In Google Chrome on Linux, the memory usage for the barcodesd180.csv and CPdata180.csv files from the human heart dataset seems to be reduced by 20% and 40%, respectively :)

Note: Quadtrees are currently disabled, because d3.quadtree() expects input in AoS format and only stores a reference to each datapoint when it generates the tree. We could store indices instead, if we first extract (x, y, index) tuples from the data and provide that as input to the function.

References:
[1] https://v8.dev/blog/elements-kinds