Use Papa Parse to enable loading of larger CSV files
TissUUmaps is now able to load and display a 2.6GB CSV dataset with 14.8 million markers, without crashing the browser or returning a request error, in Chrome and Firefox (tested on Linux). In Chrome, there seems to be a limit how large normal arrays can be, so when creating the vertex data we have to use a single large typed array instead.