Add code for creating quadtrees for new data layout
Each leaf node of a tree now stores an index instead of a reference to a marker object. Memory wise, storing a full quadtree with a leaf node for every point is very inefficient for larger datasets (more than a few million points), so they are disabled by default for now.