Add support for resolution scaling of markers
This adds a new parameter glUtils._resolutionScale that can be used to force a lower internal rendering resolution, which can improve performance. An actual scaling factor, glUtils._resolutionScaleActual, is also computed from this value and the OSD viewport size, to fix the problem with misalignment between the WebGL markers and the OSD image layers when the WebGL canvas resolution become too large (larger than 4096 pixels).