Add option for drawing markers as instanced billboards
Drawing the markers using instancing and gl.TRIANGLE_STRIP removes the size limitation that drawing as gl.POINTS has on some GPUs and WebGL implementations. The option glUtils._useInstancing can be used to switch between the two draw modes. We still probably want to clamp the billboard size, so that markers do not become too large, but this is now handled via a uniform sent to the vertex shader.